+++ CHANGES.rst
@@ -12,3 +12,6 @@ CHANGES
   use `app.add_subapp()` instead #1592
 
 - Dropped: `Application.finish()` and `Application.register_on_finish()`  #1602
+
+- Dropped: aiohttp.get(), aiohttp.options(), aiohttp.head(), aiohttp.post(),
+  aiohttp.put(), aiohttp.patch(), aiohttp.delete(), and aiohttp.ws_connect() #1593